requirements.txt
"fastapi[all]"==0.7.0
eth-brownie==1.17.0
vyper==0.3.0